Stat System and bug fixes
Added a stat system script and shifted some of the movement and shooting
variables to this script. This script will be common for all player and
enemy classes and will hold all the properties/variables and functions
that can change and are common between all of these classes. Also handled the self
shooting issue. Each bullet now knows who was its shooter and will not
harm the person who shot the bullet.