added enemy spawns and behaviors
Created type-1 enemy prefab along with their spawner. These enemies spawn
when their spawner is visible in the camera. Only works if a sprite
renderer is attached to the spawner, may fix that in future. The enemy
aims and shoots at the player when the player gets within a certain
range, and has no movement script attached to it